BIG UPDATE::
Fixed issues with the backend and front end not communicating with the database properly, fixed the delete button not working in the admin tab. Fixed the listings not properly showing up, changed the delete button to red in the manage tab, Added new routers for admin capabilities.

Claim Listing button works in the API but does not function properly inside of the front end, Please look into this frontend devs